transporting logic to new class
